api/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
docker==7.0.0
paramiko==3.4.0
